Seprated componnets into diff files
